+++ atcoder-problems-backend/src/lib.rs
@@ -1,4 +1,5 @@
 #[macro_use]
 extern crate diesel;
 
-pub(crate) mod sql;
+pub mod scraper;
+pub mod sql;
