--- atcoder-problems-scraper/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "atcoder-problems-scraper"
-version = "0.1.0"
-authors = ["Kenko Nakamura <kenkou.n@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-scraper = "0.9"
-reqwest = { version = "0.9", features = ["rustls-tls"], default-features = false }
-serde = { version = "1.0", features = ["derive"] }
-chrono = "0.4"
-regex = "1"
-env_logger = "0.6.1"
-log = "0.4.6"
-diesel = { version = "1.4.2", features = ["postgres"] }
-
-atcoder-problems-sql-common = { path = "../atcoder-problems-sql-common" }
