--- atcoder-problems-scraper/src/lib.rs
@@ -1,5 +1,0 @@
-pub mod scraper;
-
-pub use atcoder_problems_sql_common::models::{Contest, Performance, Problem, Submission};
-pub use atcoder_problems_sql_common::schema;
-pub use atcoder_problems_sql_common::sql;
